Convert AppInfo screen to TypeScript
