+++ src/components/AppInfo/RotatingIcon.tsx
@@ -1,0 +1,45 @@
+import React from 'react';
+import { TouchableWithoutFeedback } from 'react-native';
+import { Image } from 'expo-image';
+import analytics from '@react-native-firebase/analytics';
+
+import Animated, {
+    FlipInEasyX,
+    useAnimatedStyle,
+    useSharedValue,
+    withTiming
+} from 'react-native-reanimated';
+
+const RotatingIcon: React.FunctionComponent = ({ }) => {
+    const rotation = useSharedValue(0);
+    const rotationCount = useSharedValue(1);
+    const animatedStyles = useAnimatedStyle(() => {
+        return {
+            transform: [
+                { rotate: rotation.value + 'deg' },
+            ],
+        };
+    });
+
+    return <TouchableWithoutFeedback onPress={async () => {
+        rotationCount.value = rotationCount.value + 1;
+        rotation.value = withTiming((rotationCount.value * 90), { duration: 1000 });
+
+        await analytics().logEvent('app_icon');
+    }}>
+        <Animated.View style={[animatedStyles]} entering={FlipInEasyX.delay(0).duration(1000)}>
+            <Image source={require('../../../assets/icon.png')}
+                contentFit='contain'
+                style={{
+                    alignSelf: 'center',
+                    height: 100,
+                    width: 100,
+                    margin: 10,
+                    borderRadius: 20,
+                }}
+            />
+        </Animated.View>
+    </TouchableWithoutFeedback>
+}
+
+export default RotatingIcon;
